app/operational-intelligence/orchestrator-agent/requirements.txt
google-adk>=1.0.0
google-genai>=1.0.0
google-cloud-pubsub>=2.18.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
